Allowed past waiting for cash reservations to be confirmed (#360)
Reservations that are waiting for cash payment and are in the past were not allowed to be approved later because making changes to past reservations is not allowed. This change allows reservations in waiting for cash state to be approved/confirmed by staff even if the reservations are in the past.